requirements.txt
# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# gRPC
grpcio>=1.59.3
grpcio-tools>=1.59.3
protobuf>=4.25.0

# ML Libraries
# Note: For Python 3.14, use newer versions
scikit-learn>=1.5.0
numpy>=1.26.0
pandas>=2.2.0
joblib>=1.3.2

# Dashboard
streamlit>=1.28.1
requests>=2.31.0

# DVC
dvc[s3]==3.64.0
dvc-s3==3.2.1

# ClearML
clearml==1.14.0

# S3/MinIO
boto3==1.29.7

# Utilities
python-multipart==0.0.6
python-json-logger==2.0.7
pyyaml==6.0.1



